lat lon datatype in coord_netecdf.c changed to double

+++ include/cdf.h
@@ -164,8 +164,8 @@ extern Bool readintdata_netcdf(const Climatefile *,int *,const Cell *,
 extern Bool readshortdata_netcdf(const Climatefile *,short *,const Cell *,
                                  int,const Config *);
 extern Coord_netcdf opencoord_netcdf(const char *,const char *,Bool);
-extern const float *getlon_netcdf(Coord_netcdf,int *);
-extern const float *getlat_netcdf(Coord_netcdf,int *);
+extern const double *getlon_netcdf(Coord_netcdf,int *);
+extern const double *getlat_netcdf(Coord_netcdf,int *);
 extern void closecoord_netcdf(Coord_netcdf);
 extern Bool seekcoord_netcdf(Coord_netcdf,int);
 extern Bool readcoord_netcdf(Coord_netcdf,Coord *,const Coord *,unsigned int *);
